+++ src/updater/tera.rs
@@ -164,7 +164,7 @@ impl Updater for Tera {
 }
 
 impl Tera {
-    async fn engine(&self) -> tera::Tera {
+    pub async fn engine(&self) -> tera::Tera {
         let arc = ENGINE.get_or_init(|| Arc::new(Mutex::new(None)));
 
         if let Some(engine) = arc.lock().await.clone() {
@@ -180,18 +180,19 @@ impl Tera {
         engine.register_filter("replace_key", filters::object::replace_key);
         engine.register_filter("replace_value", filters::object::replace_value);
         engine.register_function("uuid_v4", function::string::uuid_v4);
-        engine.register_function("base64_encode", function::string::base64_encode);
         engine.register_filter("base64_encode", filters::string::base64_encode);
-        engine.register_function("base64_decode", function::string::base64_decode);
         engine.register_filter("base64_decode", filters::string::base64_decode);
         engine.register_filter("search", filters::object::search);
         engine.register_filter("env", filters::string::set_env);
         engine.register_function("env", function::string::env);
         engine.register_function("get_env", function::string::env);
-        engine.register_function("find", function::string::find);
         engine.register_filter("find", filters::string::find);
-        engine.register_function("extract", function::object::extract);
         engine.register_filter("extract", filters::object::extract);
+        engine.register_filter("values", filters::object::values);
+        engine.register_filter("keys", filters::object::keys);
+        engine.register_filter("update", filters::object::update);
+        engine.register_filter("map", filters::object::map);
+
         // faker
         engine.register_function("fake_words", function::faker::words);
         engine.register_function("fake_sentences", function::faker::sentences);